pyproject.toml
[tool.poetry]
name = "negot"
version = "0.1.0"
description = "Negotiation roleplay + coaching prototype (v0.1)"
authors = ["negot Contributors <dev@local>"]
license = "Proprietary"
readme = "README.md"
packages = [{ include = "negot", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"

# API + validation
fastapi = ">=0.110,<1.0"
uvicorn = {version = ">=0.24,<1.0", extras = ["standard"]}
pydantic = ">=2.7,<3.0"
pydantic-settings = ">=2.2,<3.0"
httpx = ">=0.27,<1.0"
python-dotenv = ">=1.0,<2.0"
tenacity = ">=8.2,<10.0"
orjson = ">=3.9,<4.0"

# Data
sqlalchemy = ">=2.0,<3.0"
alembic = ">=1.13,<2.0"
asyncpg = ">=0.29,<1.0"
pg8000 = ">=1.30,<2.0"

# Cost control
slowapi = ">=0.1.9,<1.0"
aiosqlite = "^0.22.0"

[tool.poetry.group.llm.dependencies]
litellm = ">=1.30,<2.0"

[tool.poetry.group.agents.dependencies]
langgraph = ">=0.1,<1.0"
instructor = ">=1.5,<2.0"

[tool.poetry.group.search.dependencies]
tavily-python = ">=0.3,<1.0"

[tool.poetry.group.observability.dependencies]
langfuse = ">=2.0,<3.0"

[tool.poetry.group.ui.dependencies]
streamlit = ">=1.33,<2.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0,<9.0"
pytest-asyncio = ">=0.23,<1.0"
ruff = ">=0.5,<1.0"
mypy = ">=1.10,<2.0"
pre-commit = ">=3.7,<5.0"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
